Restore leniency when parsing nested elements
Prior to e0140ba322c1fb02a9f43078290195c6ba882a66, elements with
children and a name other than `property` were ignored. We have code
that relies on this to declare additional configuration for the
maven-surefire-plugin that is later read from the `XppDom` object. This
commit restores the prior leniency for nested elements and ignores all
elements with children that don't have both a name and a value child
element.

Signed-off-by: Marc Philipp <marc@gradle.com>